+++ src/Products/PloneMeeting/events.py
@@ -1218,6 +1218,8 @@ def _notifyContainerModified(child):
            not IMeetingContent.providedBy(container)):
         notifyModifiedAndReindex(container)
         container = container.aq_parent
+        if container.portal_type == 'ToolPloneMeeting':
+            container.invalidateAllCache()
 
 
 def onConfigOrPloneElementAdded(element, event):